Adding main files HTML, CSS and JS
Main files added not including photos. Website does needs to some minor work. There are rendering problems in regard to zooming of page. Printing layout needs to be designed. Code in general needs to be dried out. Resumes need to be added to Developers section. Once on a sever, PHP needs to be added to the contact us sections.